Agregado btnRetry y seccion novedades a home fragment.
