+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,9 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.ONGAndroid">
+        android:theme="@style/Theme.ONGAndroid"
+        android:usesCleartextTraffic="true"
+        >
         <activity
             android:name=".view.MainActivity"
             android:exported="true">
