+++ app/src/main/java/com/melvin/ongandroid/model/apiservice/RetrofitInstance.kt
@@ -2,6 +2,7 @@ package com.melvin.ongandroid.model.apiservice
 
 //import com.melvin.ongandroid.constantsapi.ApiConstants
 
+import com.melvin.ongandroid.data.apiservice.ApiConstants
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
