+++ app/src/main/java/com/melvin/ongandroid/model/apiservice/apimodel/SlideModelResponse.kt
@@ -1,6 +1,7 @@
-package com.melvin.ongandroid.model.apimodel
+package com.melvin.ongandroid.model.apiservice.apimodel
 
 import com.google.gson.annotations.SerializedName
+import com.melvin.ongandroid.model.apimodel.ActivityModel
 
 /**
  * Data class que representa un response a consulta backend/api
