+++ app/src/main/java/com/melvin/ongandroid/viewmodel/HomeViewModel.kt
@@ -15,19 +15,6 @@ class HomeViewModel : ViewModel() {
 
     private var getActivitiesInteractor = GetActivitiesInteractor()
 
-
-    //variables livedata para cargar los datos en el carrousel de bienvenida
-    //martin re
-    private val _activities: MutableLiveData<List<CarouselItem>> = MutableLiveData()
-    val activities: LiveData<List<CarouselItem>> = _activities
-    //variables livedata para ocultar carousel en caso de que la lista de actividades
-    //este vacia
-    //martin re
-    private val _carouselIsGone: MutableLiveData<Boolean> = MutableLiveData()
-    val carouselIsgone: MutableLiveData<Boolean> = _carouselIsGone
-
-    private var getActivitiesInteractor = GetActivitiesInteractor()
-
     //variables livedata para cargar los datos en el carrousel de bienvenida
     //martin re
     private val _activities: MutableLiveData<List<CarouselItem>> = MutableLiveData()
@@ -61,5 +48,21 @@ class HomeViewModel : ViewModel() {
                 _carouselIsGone.postValue(true)
             }
         }
+    }
+
+    /*Agregado Provisional de items, para novedades, creado por Jose Luis Mora.
+    Mauro Peña.***
+     */
+
+    fun addItems(): List<CarouselItem>{
+
+        val items = mutableListOf<CarouselItem>()
+
+        items.add(CarouselItem("https://th.bing.com/th/id/OIP.7Zm8c0rmrGoURKXtK5KacwHaFj?w=225&h=180&c=7&r=0&o=5&dpr=1.25&pid=1.7"))
+        items.add(CarouselItem("https://th.bing.com/th/id/OIP.MuZj_uk2byfZ0I36Mni-HgHaD4?w=322&h=180&c=7&r=0&o=5&dpr=1.25&pid=1.7"))
+        items.add(CarouselItem("https://th.bing.com/th/id/OIP.TEbDh1Y0Zkh-PbGWC-mRrwHaD4?w=322&h=180&c=7&r=0&o=5&dpr=1.25&pid=1.7"))
+        items.add(CarouselItem("https://th.bing.com/th/id/OIP.gVwjgA2MCF8t4RExmprOHwHaEG?w=325&h=180&c=7&r=0&o=5&dpr=1.25&pid=1.7"))
+
+        return items
     }
 }
