--- app/src/main/res/layout/fragment_lastest_news_home.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/lastest_news_fragment"
-    android:orientation="vertical"
-    tools:context=".view.homefragments.LastestNewsHomeFragment">
-
-    <com.google.android.material.textview.MaterialTextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/lastest_news"
-        android:gravity="center"
-        android:layout_marginTop="10dp"
-        android:textStyle="bold"
-        android:textColor="@color/black"
-        android:textSize="25dp"/>
-
-    <org.imaginativeworld.whynotimagecarousel.ImageCarousel
-        android:id="@+id/lastest_news_carousel"
-        android:layout_width="match_parent"
-        android:layout_height="300dp"
-        android:layout_marginTop="10dp"
-        android:padding="5dp"
-        app:carouselType="SHOWCASE"
-        />
-
-
-</LinearLayout>
